+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- Options for file name casings, defaulting to `pascal` but can be set to `camel`, `snake` or `kebab`
+
 ## [0.15.0] - 2024-12-02
 
 ### Added